Reorganization, fixed pseudocode
I was originally going to have a folder for each lesson, but it looks
like that was overkill: the first lesson is now entirely contained in
one file. I'm placing notes on program useage (like OpenMP) into
appendices, because I want the main material to be about the theory of
HPC, not tools that happen to be relevant right now.

The pseudocode file had some typos and needed some additional code to be
functional and have proper highlighting. This code has been added,
though pseudocode is far from final.